+++ spring-cloud-dataflow-registry/src/main/java/org/springframework/cloud/dataflow/registry/support/DockerImage.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.cloud.dataflow.core;
+package org.springframework.cloud.dataflow.registry.support;
 
 import java.util.Arrays;
 import java.util.Collections;
